+++ backend/src/main/java/mouda/backend/auth/presentation/request/AppleUserInfoRequest.java
@@ -1,4 +1,4 @@
-package mouda.backend.auth.presentation.controller;
+package mouda.backend.auth.presentation.request;
 
 public record AppleUserInfoRequest(
 	AppleUserNameRequest name,
