+++ backend/src/main/java/mouda/backend/auth/presentation/request/AppleUserNameRequest.java
@@ -1,4 +1,4 @@
-package mouda.backend.auth.presentation.controller;
+package mouda.backend.auth.presentation.request;
 
 public record AppleUserNameRequest(
 	String firstName,
